add utils class
