+++ gamification-impl/src/main/resources/api-spec.yaml
@@ -119,8 +119,8 @@ paths:
             schema:
               $ref: '#/components/schemas/Event'
       responses:
-        '201':
-          description: created
+        '200':
+          description: Event successfully processed
   /badges:
     post:
       description: create a badge
